requirements.txt
openai
whisper
python-dotenv
pydub
markdown
pyyaml
click
